Update coordinate system for AutoOrtho/Ortho4XP compatibility
- Change DdsFilename row/col from i32 to u32 (Web Mercator indices are unsigned)
- Update filename parsing for AutoOrtho format: {row}_{col}_{maptype}{zoom}.dds
- Add map_type field to DdsFilename struct
- Add tile_to_lat_lon_center() for getting tile center coordinates
- Update TileRequest to use u32 for row/col
- Remove unnecessary negative coordinate validation
- Update COORDINATE_SYSTEM.md with comprehensive documentation

Key changes:
- Old format: +37-123_BI16.dds (signed lat/lon based)
- New format: 100000_125184_BI18.dds (unsigned Web Mercator)

Tested with real AutoOrtho scenery data from Europe, Australia, and Asia packs.

+++ xearthlayer/src/service/facade.rs
@@ -186,7 +186,7 @@ impl XEarthLayerService {
 
         // Create tile request with tile row/col coordinates and tile zoom
         // (TileGenerator will add 4 to get chunk zoom for downloads)
-        let request = TileRequest::new(tile.row as i32, tile.col as i32, tile_zoom);
+        let request = TileRequest::new(tile.row, tile.col, tile_zoom);
 
         // Generate tile
         self.generator
